NCI-Agency/anet#3172: Reorganize files, add more comments
