+++ client/src/components/Calendar/actions/index.js
@@ -1,3 +1,6 @@
+import VIEWS from "components/Calendar/utils/constants"
+import { addMonths, addYears } from "date-fns"
+
 export const ACTION_TYPES = {
   CHANGE_VIEW: 1,
   CHANGE_ACVITE_DATE: 2,
@@ -24,3 +27,27 @@ export function changeTitle(incTitle) {
 }
 
 export default ACTION_TYPES
+
+export function monthNextAction(dispatch, state) {
+  dispatch(changeViewDate(addMonths(state.viewDate, 1)))
+}
+
+export function monthPrevAction(dispatch, state) {
+  dispatch(changeViewDate(addMonths(state.viewDate, -1)))
+}
+export function yearNextAction(dispatch, state) {
+  dispatch(changeViewDate(addYears(state.viewDate, 1)))
+}
+
+export function yearPrevAction(dispatch, state) {
+  dispatch(changeViewDate(addYears(state.viewDate, -1)))
+}
+
+export const PREV_ACTIONS = {
+  [VIEWS.YEARLY]: yearPrevAction,
+  [VIEWS.MONTHLY]: monthPrevAction
+}
+export const NEXT_ACTIONS = {
+  [VIEWS.YEARLY]: yearNextAction,
+  [VIEWS.MONTHLY]: monthNextAction
+}
